requirements.txt
pymysql>=1.1.0
redis>=5.0.0
python-dotenv